+++ test_task/src/Block.js
@@ -1,51 +1,51 @@
-import React from 'react';
+import React, {useState, useContext} from 'react';
 import ReactDOM from 'react-dom';
+import {MainContext} from "./Context";
+import List from "./List";
 
 export default class Block extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
-            currentCard: null
+
         };
+
     }
 
-    dragStartHandler(e, item) {
-        console.log(item)
-        this.setState(this.currentCard = item)
+    dragStartHandler(e, item, updateData) {
+        updateData(item)
     }
 
     dragEndHandler(e) {
+        e.preventDefault()
         return undefined;
     }
 
-
-
-    render(div = <>
-        <div key={this.props.index} className="main__block border-3 rounded my-2 px-3 py-2 border d-flex"
-             draggable={true}
-             onDragStart={e => this.dragStartHandler(e, this.props.item)}
-             onDragLeave={e => this.dragEndHandler(e)}
-             onDragEnd={e => this.dragEndHandler(e)}
-
-
-        >
-            <div className="w-25">
-                <img className="img-fluid rounded-circle" src={this.props.item.picture.medium} alt=""/>
-            </div>
-            <div className="d-flex flex-column justify-content-evenly align-content-center w-75">
-                <div className="d-flex">
-                    <p className="p-0 m-0 px-2">{this.props.item.name.first}</p>
-                    <p className="p-0 m-0 px-2">{this.props.item.name.last}</p>
-                    <p className="p-0 m-0 px-2">{this.props.item.registered.date.split('T')[0].split('-').reverse().join(".")}</p>
+    render() {
+        let value = this.context;
+        return (
+            <div key={this.props.index} className="main__block border-3 rounded my-2 px-3 py-2 border d-flex"
+                 draggable={true}
+                 onDragStart={e => this.dragStartHandler(e, this.props.item, value.updateData)}
+                 onDragLeave={e => this.dragEndHandler(e)}
+                 onDragEnd={e => this.dragEndHandler(e)}
+            >
+                <div className="w-25">
+                    <img className="img-fluid rounded-circle" src={this.props.item.picture.medium} alt=""/>
                 </div>
-                <div>
-                    <p className="p-0 m-0 px-2">{this.props.item.email}</p>
+                <div className="d-flex flex-column justify-content-evenly align-content-center w-75">
+                    <div className="d-flex">
+                        <p className="p-0 m-0 px-2">{this.props.item.name.first}</p>
+                        <p className="p-0 m-0 px-2">{this.props.item.name.last}</p>
+                        <p className="p-0 m-0 px-2">{this.props.item.registered.date.split('T')[0].split('-').reverse().join(".")}</p>
+                    </div>
+                    <div>
+                        <p className="p-0 m-0 px-2">{this.props.item.email}</p>
+                    </div>
                 </div>
             </div>
-        </div>
-    </>) {
-        return div
+        )
     }
-
-
 }
+
+Block.contextType = MainContext;
