+++ test_task/src/Context.js
@@ -1,0 +1,3 @@
+import React from "react";
+
+export const MainContext = React.createContext();
